refactor(encodeHTML): remove duplicate definition
