+++ src/renderer/src/pages/home/Messages/Blocks/MainTextBlock.tsx
@@ -5,25 +5,13 @@ import type { RootState } from '@renderer/store'
 import { selectFormattedCitationsByBlockId } from '@renderer/store/messageBlock'
 import { type Model, WebSearchSource } from '@renderer/types'
 import type { MainTextMessageBlock, Message } from '@renderer/types/newMessage'
-import { cleanMarkdownContent } from '@renderer/utils/formats'
+import { cleanMarkdownContent, encodeHTML } from '@renderer/utils/formats'
 import { Flex } from 'antd'
 import React, { useMemo } from 'react'
 import { useSelector } from 'react-redux'
 import styled from 'styled-components'
 
 import Markdown from '../../Markdown/Markdown'
-
-// HTML实体编码辅助函数
-const encodeHTML = (str: string): string => {
-  const entities: { [key: string]: string } = {
-    '&': '&amp;',
-    '<': '&lt;',
-    '>': '&gt;',
-    '"': '&quot;',
-    "'": '&apos;'
-  }
-  return str.replace(/[&<>"']/g, (match) => entities[match])
-}
 
 interface Props {
   block: MainTextMessageBlock
